Collision Detection Working
Well, not really, but in most cases it's working. Implemented by shooting a raycast to the area the player wants to move to, and if that raycast collides with anything, move the player to the collision point.

There's a bug when the player's centerpoint isn't lined up with a tile, and he will just fall through, since the raycast is shot from the center of the player. To fix, we can shoot 2 addition raycasts from the player's corners, but this still doesn't completely fix the bug (in case there are 2 small platforms between each raycast).