Implemented collision detection
Implemented the best version of collision detection I could after a lot of trial and error

Issues: 2 consecutive blocks will stop the player from moving when their on the same y coordinate. To fix this, I made the colliders for platforms slightly smaller, added a buffer to the player for whenever he would be colliding with an object to ensure he's slightly floating above the collider.

Also, added logic to tell whenever the player is grounded or not (to be used for jumping)