Updates to gravity
I updated gravity again, this time is more compatible with the way I want to implement jumping. This time, there is a gravity counteracter that is set to gravity whenever the player goes from a grounded state to a not grounded state, and it decays to 0 as the player remains in a non-grounded state. Once the player is grounded again, it is immediately set to 0.